requirements.txt
httpx==0.27.2
numpy==1.26.4
opencv-python==4.10.0.82
fastapi==0.115.2
onnxruntime-gpu==1.19.0
facenet-pytorch==2.5.3
scipy==1.13.0
uvicorn==0.34.0